Core engine updates
Some boring stuff but if your interested. I created two sets a variables
for tiles. One for the first layer and another for the second layer.
Also each tile is an object instead of a number. It has it's own x, y,
crop, image, size, and other info. Chest collisions are checked by
looping through the second array for a match with x, y and type.  saved
about 10 lines of code on the render function.